+++ Source/GateRimSG1/GR_DefOf.cs
@@ -31,7 +31,10 @@ namespace GateRimSG1
         public static IncidentDef SG1_TokraSafehouseSignal;
         public static IncidentDef SG1_TokraSafehouseLeadCache;
         public static IncidentDef SG1_TokraHiddenSafehouseWorldMarker;
+        public static IncidentDef SG1_TokraHiddenSafehouseSiteIncident;
         public static WorldObjectDef SG1_TokraHiddenSafehouseMarker;
+        public static WorldObjectDef SG1_TokraHiddenSafehouseSite;
+        public static SitePartDef SG1_TokraHiddenSafehouseSitePart;
         public static IncidentDef SG1_FreeJaffaPeacefulVisitors;
         public static IncidentDef SG1_GoauldJaffaNaturalRaid;
         public static IncidentDef SG1_GoauldQueenArrival;